pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlmodel = "^0.0.19"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
[tool.poetry.scripts]
dev= "todo.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
